% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lselect.R
\name{lselect}
\alias{lselect}
\title{Returns a list subset.}
\usage{
lselect(data, column, var_names = NULL, list_name = FALSE)
}
\arguments{
\item{data}{Data frame that you want to explore.}

\item{column}{Column name that you want to select.}

\item{list_name}{For non-list columns keep names for automic vectors. Default is FALSE.}
}
\value{
A subset of a list.
}
\description{
Returns a list subset.
}
