% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize data to quickly learn about it.}
\usage{
summarize_data(data, type = "basic")
}
\arguments{
\item{data}{Data frame that you want to explore.}

\item{type}{Default is basic. For more advanced summary do type = "advanced".}
}
\value{
A data frame
}
\description{
Summarize data to quickly learn about it.
}
