% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_importance}
\alias{summarize_importance}
\title{Summary of variable imporance.}
\usage{
summarize_importance(data, var_outcome, objective = "reg:linear")
}
\arguments{
\item{data}{Data frame that you want to explore.}

\item{var_outcome}{Outcome variable.}

\item{objective}{Either "reg:linear" or "reg:logistic".}
}
\value{
A data frame
}
\description{
Summary of variable imporance.
}
